pyproject.toml
[tool.poetry]
name = "prygmatic"
version = "0.1.0"
description = "Python API framework for vercel serverless functions"
authors = ["Dallin Romney <dallinromney@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
sanic = ">=21.3.4,<21.9"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
